extract arrayToggle and arrayRotate from Project
+ add tests
