+++ src/utils/index.js
@@ -1,2 +1,4 @@
+export { default as arrayRotate } from './arrayRotate';
 export { default as arrayShuffle } from './arrayShuffle';
+export { default as arrayToggle } from './arrayToggle';
 export { default as getDayOfYear } from './getDayOfYear';
